refactor: move useAddress hook to backup and update import path
