+++ frontend/hooks/backup/useAddress.ts
@@ -1,0 +1,19 @@
+/**
+ * @note Access address via wallet or service, relative to current chain, service id
+ */
+
+// const useAddress = () => {
+//   const { service } = useServices();
+
+//   /** agent safe multisig address */
+//   const multisigAddress =
+//     service?.chain_configs?.[CHAIN_CONFIG.OPTIMISM.chainId]?.chain_data
+//       ?.multisig;
+
+//   /** agent instance EOA address */
+//   const instanceAddress =
+//     service?.chain_configs?.[CHAIN_CONFIG.OPTIMISM.chainId]?.chain_data
+//       ?.instances?.[0];
+
+//   return { instanceAddress, multisigAddress };
+// };
